pyproject.toml
[project]
name = "kormarc-auto"
version = "0.1.0"
description = "한국 도서관용 KORMARC 자동 생성 도구 — ISBN/사진 입력으로 KOLAS·DLS 호환 .mrc 출력"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "okwhr", email = "okwhrlgma@gmail.com" }]

dependencies = [
  "pymarc>=5.3",
  "requests>=2.32",
  "anthropic>=0.40",
  "pyzbar>=0.1.9",
  "Pillow>=10.4",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "python-multipart>=0.0.20",
  "python-dotenv>=1.0",
  "hanja>=0.15",
  "opencv-python>=4.10",
  "diskcache>=5.6",      # API 응답 30일 캐시 (비용 절감)
  "tenacity>=9.0",       # 재시도 데코레이터
  "rich>=13.9",          # CLI 출력 색상·표·진행률
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.7",
  "mypy>=1.13",
]
ui = [
  "streamlit>=1.39",
]

[project.scripts]
kormarc-auto = "kormarc_auto.cli:main"
kormarc-server = "kormarc_auto.server.app:run"
kormarc-ui = "kormarc_auto.ui.streamlit_app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kormarc_auto"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
